add: if return value none -> return 0
